Refactor config load/save to file manager
